smoke test successfully loads http://ibm.com
