--- source/DasBlog.SmokeTest/IWebRunner.cs
@@ -1,7 +1,0 @@
-﻿namespace DasBlog.SmokeTest
-{
-	public interface IWebRunner
-	{
-		void RunDasBlog();
-	}
-}
