+++ source/DasBlog.SmokeTest/IWebServerRunner.cs
@@ -1,0 +1,7 @@
+﻿namespace DasBlog.SmokeTest
+{
+	public interface IWebServerRunner
+	{
+		void RunDasBlog();
+	}
+}
