--- source/DasBlog.SmokeTest/WebRunner.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace DasBlog.SmokeTest
-{
-	internal class WebRunner : IWebRunner
-	{
-		public void RunDasBlog()
-		{
-			throw new NotImplementedException();
-		}
-	}
-}
