+++ source/DasBlog.SmokeTest/WebServerRunner.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace DasBlog.SmokeTest
+{
+	internal class WebServerRunner : IWebServerRunner
+	{
+		public void RunDasBlog()
+		{
+			throw new NotImplementedException();
+		}
+	}
+}
